ELS/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.0
django-cors-headers==4.4.0
django-jazzmin==3.0.0
django-material==1.12.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
pillow==10.4.0
PyJWT==2.9.0
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
urllib3==2.2.3
